backend/requirements.txt
aiomysql==0.2.0
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
asyncpg==0.30.0
certifi==2024.12.14
click==8.1.7
coverage==7.6.9
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pydantic==2.10.3
pydantic_core==2.27.1
PyMySQL==1.1.1
pytest==8.3.4
pytest-cov==6.0.0
python-dotenv==1.0.1
python-multipart==0.0.19
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tomli==2.2.1
typing_extensions==4.12.2
uvicorn==0.32.1
